[QOL-9122] move plugin module to higher level
- It's better to put the shared code into plugin.py than into an __init__.py file,
and putting it at the higher level lets us skip a bunch of '../' paths
